LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/popup-redirect-same-origin-allow-popups.https-expected.txt


FAIL Same origin popup redirects to same-origin with same-origin-allow-popups assert_equals: opener expected "false" but got "true"
FAIL Cross origin popup redirects to same-origin with same-origin-allow-popups assert_equals: opener expected "false" but got "true"
